Create Activity #3
The problem here is pretty basic, just create a multiplication table. But I was trying to do an array version of it so I don't have to manually times them. But I can't code them yet and it took be a solid minute to just figure it out, but for now can't do it yet. So I just did a while loop.

There's a much more easier and faster way to do this, but still a beginner so...